feat: add locale middleware and resolve locale utility
- Implemented localeMiddleware to handle locale resolution for both client and server contexts.
- Introduced resolveLocale function to determine the current locale based on the environment (client/server).
- Utilized AsyncLocalStorage for managing locale state on the server side.
- Integrated paraglideMiddleware for locale extraction from web requests.
