--- app/utils/locale-middleware.tsx
@@ -1,46 +1,0 @@
-import { AsyncLocalStorage } from 'node:async_hooks'
-
-import { createMiddleware } from '@tanstack/react-start'
-import { getWebRequest } from '@tanstack/react-start/server'
-
-import { baseLocale, getLocale, type Locale, overwriteGetLocale } from '~/paraglide/runtime.js'
-import { paraglideMiddleware } from '~/paraglide/server.js'
-
-export const localeMiddleware = createMiddleware()
-  .client(async (context) => {
-    try {
-      const request = getWebRequest()
-
-      if (!request) {
-        return context.next({
-          sendContext: {
-            locale: baseLocale as Locale,
-          },
-        })
-      }
-
-      const resolvedLocale = await new Promise<Locale>((resolve) =>
-        paraglideMiddleware(request, async ({ locale }) => {
-          resolve(locale)
-        }),
-      )
-
-      return context.next({
-        sendContext: {
-          locale: resolvedLocale,
-        },
-      })
-    } catch {
-      return context.next({
-        sendContext: {
-          locale: getLocale(),
-        },
-      })
-    }
-  })
-  .server(async (context) => {
-    const storage = new AsyncLocalStorage<Locale>()
-    overwriteGetLocale(() => storage.getStore() ?? baseLocale)
-
-    return await storage.run(context.context.locale, async () => await context.next())
-  })
