+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tanstack-start-example-basic",
+  "name": "tanstack-start-example-paraglide",
   "private": true,
   "sideEffects": false,
   "type": "module",
@@ -11,24 +11,24 @@
     "test:types": "tsc --noEmit"
   },
   "dependencies": {
-    "@tanstack/react-router": "^1.114.29",
-    "@tanstack/react-start": "^1.114.30",
-    "@tanstack/router-devtools": "^1.114.29",
+    "@tanstack/react-router": "^1.115.2",
+    "@tanstack/react-router-devtools": "^1.115.2",
+    "@tanstack/react-start": "^1.115.2",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "tailwind-merge": "^2.6.0",
     "vinxi": "0.5.3"
   },
   "devDependencies": {
-    "@inlang/cli": "^3.0.9",
-    "@inlang/paraglide-js": "^2.0.8",
-    "@tailwindcss/postcss": "^4.0.17",
-    "@types/node": "^22.13.14",
-    "@types/react": "^19.0.12",
-    "@types/react-dom": "^19.0.4",
+    "@inlang/cli": "^3.0.10",
+    "@inlang/paraglide-js": "^2.0.11",
+    "@tailwindcss/postcss": "^4.1.3",
+    "@types/node": "^22.14.0",
+    "@types/react": "^19.1.0",
+    "@types/react-dom": "^19.1.2",
     "postcss": "^8.5.3",
-    "tailwindcss": "^4.0.17",
-    "typescript": "^5.8.2",
+    "tailwindcss": "^4.1.3",
+    "typescript": "^5.8.3",
     "vite-tsconfig-paths": "^5.1.4"
   }
 }
